Split the world database into several files for better organization
